% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstabs.R
\name{crosstable}
\alias{crosstable}
\title{Build an attractive cross-tabulation}
\usage{
crosstable(df, var1, var2, weight_var = NULL, margins = FALSE,
  percent = FALSE)
}
\arguments{
\item{df}{a \code{data.frame} containing the variables to cross-tabulate.}

\item{var1}{a character field name in \code{df} that will make the rows of the table.}

\item{var2}{a character field name in \code{df} that will make the columns of the table.}

\item{weight_var}{a character field name in \code{df} that weights the data. If
\code{NULL}, will treat each record equally.}

\item{margins}{boolean, should the row and column sums be appended?
default \code{FALSE}}

\item{percent}{boolean, should the table be a percentage?
default \code{FALSE}}
}
\value{
an array that can be cast with \code{kable()} or examined directly.
}
\description{
Build an attractive cross-tabulation
}
\details{
If \code{var1,var2} are numeric vectors, then the name of the field
  will be prepended to the row or column names. If not, then the bare
  attribute name will be given.
}
\examples{
crosstable(links, "facility_group", "area_name")
crosstable(links, "facility_group", "area_name", margins = TRUE)
crosstable(links, "facility_group", "area_name", percent = TRUE)
crosstable(links, "facility_group", "area_type", percent = TRUE)


}
