% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cut_abserror}
\alias{cut_abserror}
\title{Cut absolute differences into ranges}
\usage{
cut_abserror(x)
}
\arguments{
\item{x}{a vector of absolute error measurements, as in \eqn{x - y}}
}
\value{
a factor variable with each entry in x binned.
}
\description{
Cut absolute differences into ranges
}

