% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cut_diverror}
\alias{cut_diverror}
\title{Cut diverging differences into ranges}
\usage{
cut_diverror(x)
}
\arguments{
\item{x}{a vector of diverging error measurements, as in \eqn{x - y}}
}
\value{
a factor variable with each entry in x binned.
}
\description{
Cut diverging differences into ranges
}

