% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cut_volumes}
\alias{cut_volumes}
\title{Cut volumes into pretty levels}
\usage{
cut_volumes(x)
}
\arguments{
\item{x}{Volume levels}
}
\value{
A labeled factor variable of \code{length(x)} with the levels of
  \code{x} cut into bins.
}
\description{
Cut volumes into pretty levels
}

