% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flows}
\alias{flows}
\title{District-to-District Flows}
\format{A data frame with 20 observations and 3 variables.
\describe{
  \item{origin}{Origin zone}
  \item{destination}{Destination zone}
  \item{flow}{Flow from origin to destination}
}}
\usage{
flows
}
\description{
A dataset containing district-to-district flows.
}
\keyword{datasets}
