% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_tables.R
\name{link_measures_table}
\alias{link_measures_table}
\title{Link Measures Table}
\usage{
link_measures_table(links, volume, distance = NULL, speed = NULL,
  ffspeed = NULL, capacity = NULL, group_field = NULL,
  volume_breaks = c(0, 5, 10, 15, 20, 40, 60, Inf), type = c("vmt", "vht",
  "vhd", "voc"))
}
\arguments{
\item{links}{Model link table as a \code{tidy} data frame, with each row
representing an analysis link. The function assumes that the data is
already tidy: two-way links should already be converted to single values,
etc.}

\item{volume}{Character string identifying the modeled volume in the link
table.}

\item{distance}{Character string identifying the distance in the link table.}

\item{speed}{Character string identifying the modeled speed in the link table.}

\item{ffspeed}{Character string identifying the free-flow speed in the link table.}

\item{capacity}{Character string identifying the capacity in the link table.}

\item{group_field}{Character string identifying variable to
group observations by, for example facility type. If set to same value
as \code{volume}, will cut into bins.}

\item{volume_breaks}{Numeric vector passed on to \code{cut()} identifying the
breakpoints in the volume groups. Number in thousands, i.e.: 10, 20}

\item{type}{Which type of table to print. Currently supports VMT, VHT, VHD and VOC.}
}
\value{
A \code{data_frame} with the link summary table.
}
\description{
Link Measures Table
}
\examples{
link_measures_table(links, "volume", distance = "distance",
  group_field = "area_name", type = "vmt")
link_measures_table(links, "volume", distance = "distance",
  group_field = "facility_group", type = "vmt")
link_measures_table(links, "volume", distance = "distance", speed = "speed",
  group_field = "area_name", type = "vht")
link_measures_table(links, "volume", distance = "distance", speed = "speed",
  group_field = "facility_group", type = "vht")
link_measures_table(links, "volume", distance = "distance", speed = "speed",
  ffspeed = "ffspeed",  group_field = "area_name", type = "vhd")
link_measures_table(links, "volume", distance = "distance", speed = "speed",
  ffspeed = "ffspeed",  group_field = "facility_group", type = "vhd")
link_measures_table(links, "volume", capacity = "capacity",
  group_field = "area_name", type = "voc")
link_measures_table(links, "volume", capacity = "capacity",
  group_field = "facility_group", type = "voc")


}

