% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_tables.R
\name{link_stats_table}
\alias{link_stats_table}
\title{Link Statistics Table}
\usage{
link_stats_table(links, volume, count, group_field = NULL,
  volume_breaks = c(0, 5, 10, 15, 20, 40, 60, Inf), type = c("rmse",
  "flow"))
}
\arguments{
\item{links}{Model link table as a \code{tidy} data frame, with each row
representing an analysis link. The function assumes that the data is
already tidy: two-way links should already be converted to single values,
etc.}

\item{volume}{Character string identifying the modeled volume in the link
table.}

\item{count}{Character string identifying the calibration/validation counts
in the link table.}

\item{group_field}{Character string identifying variable to
group observations by, for example facility type. If set to same value
as \code{volume}, will cut into bins.}

\item{volume_breaks}{Numeric vector passed on to \code{cut()} identifying the
breakpoints in the volume groups. Number in thousands, i.e.: 10, 20}

\item{type}{Which type of table to print. Currently supports percent RMSE and
total flow deviation.}
}
\value{
A \code{data_frame} with the link summary table.
}
\description{
Link Statistics Table
}
\examples{
link_stats_table(links, "volume", "count", group_field = "area_name", type = "rmse")
link_stats_table(links, "volume", "count", group_field = "area_name", type = "flow")
link_stats_table(links, "volume", "count", group_field = "facility_group", type = "rmse")
link_stats_table(links, "volume", "count", group_field = "facility_group", type = "flow")




}

