% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_targets.R
\name{link_test}
\alias{link_test}
\title{Construct a test call}
\usage{
link_test(d, crit, string, target)
}
\arguments{
\item{d}{A links data_frame with the universe of the test}

\item{crit}{A quoted evaluationa call describing a passed test.}

\item{string}{A character string defining the test}

\item{target}{The target passing value}
}
\value{
A \code{data_frame} with a description and result of the test.
}
\description{
Construct a test call
}
\examples{
\dontrun{
# This is a pretend test to illustrate
link_test(links, quote(volume > count), "Volume over count", 50)
}
}

