% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mdd}
\alias{mdd}
\title{Maximum desirable deviation curve}
\format{A data frame with 7 rows and 2 variables:
\describe{
  \item{volume}{the volume category of the road}
  \item{mdd}{maximum desirable deviation, in percent}
}}
\usage{
mdd
}
\description{
A dataset containing the maximum desirable error for links by volume.
}
\keyword{datasets}
