% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mdd.R
\name{mdd_table}
\alias{mdd_table}
\title{Create mdd table}
\usage{
mdd_table(from = 1000, to = 151000, step = 10000)
}
\arguments{
\item{from}{Where the table should start.}

\item{to}{the maximum value of the table.}

\item{step}{the step size for \code{seq()}.}
}
\value{
a data frame containing the maximum desirable deviation for a range
  of volumes.
}
\description{
Creates a data frame containing the maximum desirable deviation
  for a range of volumes using the guidance provided in NCHRP report 255
  (updated by NCHRP 765). The MDD curve is a step-wise function that accepts
  larger deviations on roads with lower volume.
}
\examples{
mdd_table()
mdd_table(from = 10000, to = 100000, step = 20000)

}
