% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{numerify}
\alias{numerify}
\title{Numerify a mixed vector}
\usage{
numerify(x)
}
\arguments{
\item{x}{A mixed vector consisting of items such as `person1` and `3auto`}
}
\description{
Numerify a mixed vector
}
\details{
Serves as a wrapper to `gsub` and `type.convert`
}
\examples{
x <- c("person1", "person2", "person3")
numerify(x)

}
