% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/outviz.R
\docType{package}
\name{outviz}
\alias{outviz}
\alias{outviz-package}
\alias{outviz}
\alias{outviz-package}
\title{outviz: A package for standardized model outputs}
\description{
outviz: A package for standardized model outputs

This package reduces repetitive table generation and validation plot outputs
to easy to use functions that can be included in interactive documents.
}
