% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pct_error}
\alias{pct_error}
\title{Calculate percent error}
\usage{
pct_error(x, y)
}
\arguments{
\item{x}{Comparison value}

\item{y}{Reference value}
}
\value{
a numeric vector of \code{length(x)} containing the percent deviation
  from \code{y}
}
\description{
Calculate percent error
}
