% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pct_rmse}
\alias{pct_rmse}
\title{Calculate percent RMSE}
\usage{
pct_rmse(x, y)
}
\arguments{
\item{x}{Comparison value}

\item{y}{Reference value}
}
\value{
The percent root mean squared error between \code{x} and \code{y}
}
\description{
Calculate percent RMSE
}
\details{
\eqn{\sqrt{\sum(x - y)^2/(n - 1)} / \bar{y} * 100}
}
\examples{
x <- runif(10)
y <- x + rnorm(10)
pct_rmse(x, y)

}
\seealso{
rmse
}

