% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlfd.R
\name{plot_tlfd}
\alias{plot_tlfd}
\title{Plot a trip length frequency distribution comparison}
\usage{
plot_tlfd(model, target, names, xaxis, yaxis)
}
\arguments{
\item{model}{A dataframe with the following fields
\describe{
  \item{bin}{A column listing the bin values}
  \item{count}{A column listing the count per bin}
}}

\item{target}{A dataframe with the following fields
\describe{
  \item{bin}{A column listing the bin values}
  \item{count}{A column listing the count per bin}
}}

\item{names}{Character vector of names for traces}

\item{xaxis}{Name to plot on x axis}

\item{yaxis}{Name to plot on y axis}
}
\value{
A ggplot2 object
}
\description{
Plot a trip length frequency distribution comparison
}
