% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mdd.R
\name{plotly_mdd}
\alias{plotly_mdd}
\title{Maximum desirable devation plotly object}
\usage{
plotly_mdd(links, volume, count, color_field, id = NULL)
}
\arguments{
\item{links}{Model link table as a \code{tidy} data frame, with each row
representing an analysis link. The function assumes that the data is
already tidy: two-way links should already be converted to single values,
etc.}

\item{volume}{Character string identifying the modeled volume in the link
table.}

\item{count}{Character string identifying the calibration/validation counts
in the link table.}

\item{color_field}{(Optional) character string identifying variable to color
observations by, for example facility type.}

\item{id}{A link identification field. If NULL, then use \code{rownames(links)}}
}
\value{
a plotly object
}
\description{
Maximum desirable devation plotly object
}
\examples{
plotly_mdd(links, "volume", "count", "facility_group")

}

