% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlfd.R
\name{plotly_tlfd}
\alias{plotly_tlfd}
\title{Plot a trip length frequency distribution comparison as an interactive figure}
\usage{
plotly_tlfd(model, target)
}
\arguments{
\item{model}{A dataframe with a distance field and a frequency field representing
the model or current scenario.}

\item{target}{A dataframe with a distance field and a frequency field representing
the target or reference scenario.}
}
\value{
A plotly object
}
\description{
Plot a trip length frequency distribution comparison as an interactive figure
}

