% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_tables.R
\name{refactor_levels}
\alias{refactor_levels}
\title{Refactor a variable so that it prints properly.}
\usage{
refactor_levels(df, group_field = NULL)
}
\arguments{
\item{df}{A data_frame}

\item{group_field}{a character name of a variable in DF to refactor}
}
\value{
A data_frame with the appropriate factoring
}
\description{
Refactor a variable so that it prints properly.
}
\details{
If the variable is already a named factor, then this simply returns
}

