% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_smooth_func.R
\docType{data}
\name{StatSmoothFunc}
\alias{StatSmoothFunc}
\title{Stat Smoothing and Labeling for ggplot2 objects}
\format{An object of class \code{StatSmooth} (inherits from \code{Stat}, \code{ggproto}) of length 4.}
\usage{
StatSmoothFunc
}
\arguments{
\item{...}{Arguments received from stat_smooth_func}
}
\description{
Stat Smoothing and Labeling for ggplot2 objects
}
\keyword{datasets}

