% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_tables.R
\name{volume_levels}
\alias{volume_levels}
\title{Cut a volumes variable into levels}
\usage{
volume_levels(df, volume_variable, breaks)
}
\arguments{
\item{df}{A data_frame}

\item{volume_variable}{a numeric vector giving the modeled link volume}

\item{breaks}{breakpoints for the volumes vector in thousands}
}
\value{
A data_frame with the volume variable recoded into factors
}
\description{
Cut a volumes variable into levels
}

