% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cut_diverror}
\alias{cut_diverror}
\title{Cut diverging percent error into ranges}
\usage{
cut_diverror(x)
}
\arguments{
\item{x}{a vector of percent error measurements, as from \link{pct_error}}
}
\value{
a factor variable with each entry in x binned.
}
\description{
Cut diverging percent error into ranges
}

