% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cut_error}
\alias{cut_error}
\title{Cut percent error into ranges}
\usage{
cut_error(x)
}
\arguments{
\item{x}{a vector of error measurements}
}
\value{
a factor showing the bin
}
\description{
Cut percent error into ranges
}

