% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{links}
\alias{links}
\title{Link output table}
\format{A data frame with 505 observations and 8 variables.
\describe{
  \item{a, b}{Node identifiers}
  \item{facility_type}{The facility type of the links}
  \item{facility_group}{High-level facility types}
  \item{area_type}{The area type of the links}
  \item{area_name}{The area type of the links with named description}
  \item{volume}{Modeled link volume}
  \item{screenline}{The screenline that the links belong to}
  \item{count}{The calibration year count value}
  \item{distance}{The distance value in miles }
}}
\usage{
links
}
\description{
A dataset containing example link volume and count information to demonstrate
package functionality.
}
\keyword{datasets}

