% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mdd.R
\name{plot_mdd}
\alias{plot_mdd}
\title{Maximum desirable deviation plot}
\usage{
plot_mdd(links, volume, count, color_field = NULL)
}
\arguments{
\item{links}{Model link table as a \code{tidy} data frame, with each row
representing a single direction of a single link.}

\item{volume}{Character string identifying the modeled volume in the link
table.}

\item{count}{Character string identifying the calibration/validation counts
in the link table.}

\item{color_field}{(Optional) character string identifying variable to color
observations by, for example facility type.}
}
\value{
A ggplot2 plot object
}
\description{
Maximum desirable deviation plot
}
\examples{
plot_mdd(links, "volume", "count")
plot_mdd(links, "volume", "count", color_field = "facility_group")

}

