% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_validation.R
\name{plot_validation}
\alias{plot_validation}
\title{Link validation plot}
\usage{
plot_validation(links, volume, count, color_field = NULL, show_lm = FALSE)
}
\arguments{
\item{links}{Model link table as a \code{tidy} data frame, with each row
representing an analysis link. The function assumes that the data is
already tidy: two-way links should already be converted to single values,
etc.}

\item{volume}{Character string identifying the modeled volume in the link
table.}

\item{count}{Character string identifying the calibration/validation counts
in the link table.}

\item{color_field}{(Optional) character string identifying variable to color
observations by, for example facility type.}

\item{show_lm}{If \code{TRUE}, plot a linear model regression fit alongside
the fitted model's equation. Default \code{FALSE} will print spline
regression fit without any equation.}
}
\value{
A ggplot2 plot object.
}
\description{
Link validation plot
}
\examples{
plot_validation(links, "volume", "count", show_lm = TRUE)
plot_validation(links, "volume", "count", color_field = "facility_group")

}

