% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlfd.R
\name{prep_tlfd_data}
\alias{prep_tlfd_data}
\title{Prepares model data for tlfd plotting. Given a skim table and model trip
table, will return a table in the format needed by \code{plotly_tlfd} and
\code{plot_tlfd}.}
\usage{
prep_tlfd_data(skim, model, max_dist = 60, pct = TRUE)
}
\arguments{
\item{skim}{An impedance dataframe with the following columns
\describe{
  \item{from}{From TAZ}
  \item{to}{To TAZ}
  \item{imp}{Impendance between from and to TAZs
  }
}}

\item{model}{A trip dataframe with the following columns.
\describe{
  \item{from}{From TAZ}
  \item{to}{To TAZ}
  \item{trips}{Number of trips between from and to TAZs}
}}

\item{max_dist}{Maximum distance bin to create}

\item{pct}{\code{TRUE/FALSE} If true, a percentage distribution will be
returned. If false, the raw counts will be returned.}
}
\value{
A dataframe with a \code{bin} column of impedance and a \code{count}
  column of observations.
}
\description{
Prepares model data for tlfd plotting. Given a skim table and model trip
table, will return a table in the format needed by \code{plotly_tlfd} and
\code{plot_tlfd}.
}
