% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rmse}
\alias{rmse}
\title{Calculate RMSE}
\usage{
rmse(x, y)
}
\arguments{
\item{x}{Comparison value}

\item{y}{Reference value}
}
\value{
The root mean squared error between \code{x} and \code{y}
}
\description{
Calculate RMSE
}
\details{
\eqn{\sqrt{\sum(x - y)^2/(n - 1)}}
}
\examples{
x <- runif(10)
y <- x + rnorm(10)
rmse(x, y)

}

