% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_tables.R
\name{table_rmse}
\alias{table_rmse}
\title{Percent RMSE table}
\usage{
table_rmse(links, volume, count, group_field = NULL)
}
\arguments{
\item{links}{Model link table as a \code{tidy} data frame, with each row
representing a single direction of a single link.}

\item{volume}{Character string identifying the modeled volume in the link
table.}

\item{count}{Character string identifying the calibration/validation counts
in the link table.}

\item{group_field}{Character string identifying variable to
group observations by, for example facility type. If set to same value
as \code{volume}, will cut into bins.}
}
\value{
A \code{data_frame} with the link summary table.
}
\description{
Percent RMSE table
}
\examples{
table_rmse(links, "volume", "count")

a <- table_rmse(links, "volume", "count", group_field = "facility_group")
knitr::kable(a)

}

